Version 1.0 mostly ready
- Dramatically scaled down core edits to the essentials;
- Core edits now affect and patch two core functions: redirect() and
inline_error(). An inline success message is generated everytime the
redirect() function is called, setting the global $errors variable as
filled in with the parsed message from inline_success() function. This
has 3 main positive aspects: it saves template edits, it saves a lot of
work (it's a sort of "global" solution instead of a "local" one), and
it allows the data to be available instantly instead of old data
populating the inputs when no errors are encountered (basically, it
tricks MyBB into checking for errors, but instead there's a success
message).
- The inline_error() patch ensures no errors are displayed at the same
time of a success message.
- The redirect() patch checks for "do_" substring in
$mybb->input['action'] field and removes it to display the success
message with the new data filled in as explained above.
- A master switch setting has been added to let admins enable and
disable the plugin effects with ease.
- This new version is poorly tested but works fine as for the first
tests. Feel free to install and check it out!